refactor(break): Switch full login flow out with Auth0
