+++ pages/_app.tsx
@@ -1,4 +1,4 @@
-import { Auth0Provider, useAuth0 } from '@auth0/auth0-react';
+import { Auth0Provider } from '@auth0/auth0-react';
 import 'compiled.css';
 import Router from 'next/router';
 import Footer from 'features/Footer';
@@ -6,7 +6,7 @@ import Header from 'features/Header';
 import Settings from 'features/Settings';
 import type { AppProps } from 'next/app';
 import Head from 'next/head';
-import React, { useEffect } from 'react';
+import React from 'react';
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 const onRedirectCallback = (appState: any) => {
@@ -47,12 +47,11 @@ const Layout: React.FC<AppProps> = ({ Component, pageProps }: AppProps) => (
 						? process.env.NEXT_PUBLIC_REDIRECT_URI
 						: undefined,
 				audience: process.env.NEXT_PUBLIC_AUDIENCE,
-				scope: 'account:read',
+				scope: 'account:read profile',
 			}}
 		>
 			<Settings>
 				<div className="flex min-h-screen flex-col">
-					<Auth />
 					<Header />
 					<main className="h-full grow bg-white text-gray-900 dark:bg-gray-900 dark:text-gray-200">
 						<Component {...pageProps} />
@@ -65,83 +64,3 @@ const Layout: React.FC<AppProps> = ({ Component, pageProps }: AppProps) => (
 );
 
 export default Layout;
-
-const Auth = () => {
-	const { isAuthenticated, error, user } = useAuth0();
-
-	if (error) {
-		console.log(error);
-		return <div>Oops... {error.message}</div>;
-	}
-
-	if (!isAuthenticated) {
-		return (
-			<div>
-				<div>No user authenticated</div>
-				<LoginButton />
-			</div>
-		);
-	}
-
-	console.log(user);
-
-	return (
-		<div>
-			<UserDisplay />
-			<LogoutButton />
-			<Authorized />
-		</div>
-	);
-};
-
-const UserDisplay = () => {
-	const { user } = useAuth0();
-
-	return <div>{user?.sub}</div>;
-};
-
-const Authorized = () => {
-	const { getAccessTokenSilently, loginWithRedirect } = useAuth0();
-	useEffect(() => {
-		(async () => {
-			try {
-				const token = await getAccessTokenSilently();
-			} catch (e: any) {
-				if (e.error === 'login_required') {
-					loginWithRedirect();
-				}
-				if (e.error === 'consent_required') {
-					loginWithRedirect();
-				}
-				throw e;
-			}
-		})();
-	}, [getAccessTokenSilently, loginWithRedirect]);
-
-	return <></>;
-};
-
-const LoginButton = () => {
-	const { loginWithRedirect } = useAuth0();
-	return <button onClick={() => loginWithRedirect()}>Login in</button>;
-};
-
-const LogoutButton = () => {
-	const { isAuthenticated, logout } = useAuth0();
-
-	return isAuthenticated ? (
-		<button
-			onClick={() => {
-				logout({
-					logoutParams: {
-						returnTo: window.location.origin,
-					},
-				});
-			}}
-		>
-			Log out
-		</button>
-	) : (
-		<></>
-	);
-};
