+++ src/features/login/LoginState.tsx
@@ -1,38 +1,23 @@
+import { useAuth0, User } from '@auth0/auth0-react';
 import { useOnOutsideMouseDown } from '@oliverflecke/components-react';
 import React, { useCallback, useRef, useState } from 'react';
-import { User } from 'utils/githubAuth';
 import LoginButton from './LoginButton';
-import LoginMenu from './LoginMenuProps';
+import LoginMenu from './LoginMenu';
 import UserAvatar from './UserAvatar';
 
-interface LoginStateProps {
-	user: User | null;
-	authorizeUrl: string;
-	logoutUrl: string;
-}
+const LoginState: React.FC = () => {
+	const { user } = useAuth0();
 
-const LoginState: React.FC<LoginStateProps> = ({
-	user,
-	authorizeUrl,
-	logoutUrl,
-}) =>
-	user === null ? (
-		<LoginButton authorizeUrl={authorizeUrl} />
-	) : (
-		<LoginDropDownMenu user={user} logoutUrl={logoutUrl} />
-	);
+	return user ? <LoginDropDownMenu user={user} /> : <LoginButton />;
+};
 
 export default LoginState;
 
 interface LoginDropDownMenuProps {
 	user: User;
-	logoutUrl: string;
 }
 
-const LoginDropDownMenu: React.FC<LoginDropDownMenuProps> = ({
-	user,
-	logoutUrl,
-}) => {
+const LoginDropDownMenu: React.FC<LoginDropDownMenuProps> = ({ user }) => {
 	const [isOpen, setIsOpen] = useState(false);
 
 	const ref = useRef<HTMLDivElement>(null);
@@ -43,13 +28,12 @@ const LoginDropDownMenu: React.FC<LoginDropDownMenuProps> = ({
 
 	return (
 		<div ref={ref} className="relative flex items-center space-x-4">
-			<span className="hidden sm:inline">{user.login}</span>
 			<div className="group">
 				<button onClick={() => setIsOpen(x => !x)}>
-					<UserAvatar user={user} />
+					<UserAvatar pictureUrl={user.picture} />
 				</button>
 			</div>
-			<LoginMenu isOpen={isOpen} logoutUrl={logoutUrl} />
+			<LoginMenu isOpen={isOpen} />
 		</div>
 	);
 };
