+++ src/features/login/UserAvatar.tsx
@@ -1,14 +1,13 @@
 import React from 'react';
-import { User } from 'utils/githubAuth';
 import Image from 'next/image';
 
 interface UserAvatarProps {
-	user: User;
+	pictureUrl?: string;
 }
 
-const UserAvatar: React.FC<UserAvatarProps> = ({ user }) => (
+const UserAvatar: React.FC<UserAvatarProps> = ({ pictureUrl }) => (
 	<Image
-		src={`${user.avatar_url}&s=80`}
+		src={pictureUrl ?? ''}
 		width={40}
 		height={40}
 		alt="Avatar of the logged in user"
